fix(lab): PR-AUDIT-L-H-batch-a — High priority fixes (6 issues)
L-H-1: Замена native confirm() на useConfirm() в LabTemplateWorkbench
  - handleArchiveTemplate теперь использует portal-dialog с focus-trap
  - Согласованность с LabReportWorkbench (Finalize, Revise — useConfirm)

L-H-2: Удаление мёртвого markReady кода
  - api/labReporting.js: markReady() method удалён
  - utils/labReportActions.js: mark_ready удалён из LAB_REPORT_ACTION_CAN_FIELD
  - WF-round5 убрал UI-кнопку — endpoint больше не нужен

L-H-3: Унификация маппинга статусов в labStatusConfig.js
  - Создан utils/labStatusConfig.js — единый источник истины
  - LAB_REPORT_STATUS_CONFIG + LAB_QUEUE_STATUS_CONFIG + getLabStatusConfig()
  - labUiLabels.js: formatLabStatus/getLabStatusVariant делегируют в config
  - LabStatusStepper.jsx: использует LAB_REPORT_STATUS_CONFIG из config
  - Устранено дублирование в 3 файлах

L-H-4: Миграция inline-стилей LabQueueWorkbench в CSS
  - lab.css: добавлено ~280 строк классов .lqw-* и .lrw-*
  - LabQueueWorkbench.jsx: убраны inline-стили (~150 строк)
  - Skeleton, search, status-filter, queue-card — теперь CSS-классы

L-H-5: Responsive grid для полей в LabReportWorkbench
  - lrw-field-row класс с @media (max-width: 1024px) и 720px fallback
  - Tablet: 3 колонки, Mobile: вертикальный stack

L-H-6: Декомпозиция LabTemplateWorkbench (1598 → ~600 строк)
  - templateEditor/config.js: shared константы
  - templateEditor/utils.js: hydrateVersion, buildVersionPayload, etc.
  - templateEditor/NewTemplateDialog.jsx
  - templateEditor/ReferenceRuleEditor.jsx (~290 строк)
  - templateEditor/ContentTab.jsx (~250 строк)
  - templateEditor/DesignTab.jsx (~50 строк)
  - templateEditor/SignersTab.jsx (~25 строк)
  - templateEditor/PreviewTab.jsx (~70 строк)
  - Контракт-тесты обновлены под новую структуру

Дополнительно (из Batch A):
  - L-M-3 fix: убрана вложенная Button внутри role="button"
  - L-M-4 fix: skeleton уважает prefers-reduced-motion через CSS @media
  - L-M-5 fix: PII details используют CSS-класс вместо ::-webkit-details-marker хака
